# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
  dto.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/dto.ApiError'
      meta:
        $ref: '#/definitions/dto.ResponseMeta'
      success:
        type: boolean
    type: object
...
  dto.ChainAnalysis:
    properties:
      criticalPathPosition:
        type: string
      floatHours:
        type: integer
      longestChain:
        type: integer
    type: object
...
  dto.CreateDependencyRequest:
    properties:
      dependencyType:
        enum:
        - finish_to_start
        - start_to_start
        - finish_to_finish
        - start_to_finish
        type: string
      dependsOnTaskId:
        type: string
      lagHours:
        minimum: 0
        type: integer
      taskId:
        type: string
    required:
    - dependencyType
    - dependsOnTaskId
    - taskId
    type: object
  dto.CreateDependencyResponse:
    properties:
      createdAt:
        type: string
      dependencyId:
        type: string
      dependencyType:
        type: string
      dependsOnTaskId:
        type: string
      impact:
        $ref: '#/definitions/dto.DependencyImpact'
      lagHours:
        type: integer
      taskId:
        type: string
      validation:
        $ref: '#/definitions/dto.DependencyValidation'
    type: object
...
  dto.CriticalPathInfo:
    properties:
      taskIds:
        items:
          type: string
        type: array
      tasks:
        items:
          $ref: '#/definitions/dto.CriticalPathTask'
        type: array
      totalDuration:
        type: integer
    type: object
  dto.CriticalPathResponse:
    properties:
      calculatedAt:
        type: string
      criticalPath:
        $ref: '#/definitions/dto.CriticalPathInfo'
      nonCriticalTasks:
        items:
          $ref: '#/definitions/dto.NonCriticalTask'
        type: array
      projectDuration:
        type: integer
      projectId:
        type: string
    type: object
  dto.CriticalPathTask:
    properties:
      earliestFinish:
        type: string
      earliestStart:
        type: string
      estimatedHours:
        type: integer
      floatHours:
        type: integer
      latestFinish:
        type: string
      latestStart:
        type: string
      taskId:
        type: string
      title:
        type: string
    type: object
...
  dto.DependencyGraphEdge:
    properties:
      id:
        type: string
      lagHours:
        type: integer
      source:
        type: string
      target:
        type: string
      type:
        type: string
    type: object
  dto.DependencyGraphNode:
    properties:
      id:
        type: string
      status:
        type: string
      title:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
  dto.DependencyGraphResponse:
    properties:
      edges:
        items:
          $ref: '#/definitions/dto.DependencyGraphEdge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/dto.DependencyGraphNode'
        type: array
      projectId:
        type: string
    type: object
  dto.DependencyImpact:
    properties:
      affectedTasks:
        items:
          type: string
        type: array
      criticalPathChanged:
        type: boolean
      newProjectEndDate:
        type: string
    type: object
  dto.DependencyInfo:
    properties:
      dependencyId:
        type: string
      dependencyType:
        type: string
      dependsOnTaskId:
        type: string
      isBlocking:
        type: boolean
      lagHours:
        type: integer
      status:
        type: string
    type: object
  dto.DependencySection:
    properties:
      direct:
        items:
          $ref: '#/definitions/dto.DependencyInfo'
        type: array
      indirect:
        items:
          $ref: '#/definitions/dto.IndirectDependency'
        type: array
    type: object
  dto.DependencyValidation:
    properties:
      valid:
        type: boolean
      wouldCreateCycle:
        type: boolean
    type: object
  dto.DependencyWarning:
    properties:
      message:
        type: string
      type:
        type: string
    type: object
  dto.DependentInfo:
    properties:
      dependencyType:
        type: string
      isBlocked:
        type: boolean
      taskId:
        type: string
    type: object
  dto.DependentsSection:
    properties:
      direct:
        items:
          $ref: '#/definitions/dto.DependentInfo'
        type: array
      indirect:
        items:
          $ref: '#/definitions/dto.IndirectDependency'
        type: array
    type: object
...
  dto.IndirectDependency:
    properties:
      depth:
        type: integer
      path:
        items:
          type: string
        type: array
    type: object
...
  dto.NonCriticalTask:
    properties:
      floatHours:
        type: integer
      taskId:
        type: string
      title:
        type: string
    type: object
...
  dto.ResponseMeta:
    properties:
      hasMore:
        type: boolean
      nextCursor:
        type: string
      page:
        type: integer
      perPage:
        type: integer
      requestId:
        type: string
      timestamp:
        type: string
      total:
        type: integer
    type: object
...
  dto.TaskDependenciesResponse:
    properties:
      chainAnalysis:
        $ref: '#/definitions/dto.ChainAnalysis'
      dependencies:
        $ref: '#/definitions/dto.DependencySection'
      dependents:
        $ref: '#/definitions/dto.DependentsSection'
      taskId:
        type: string
    type: object
...
  dto.ValidateDependencyRequest:
    properties:
      dependencyType:
        enum:
        - finish_to_start
        - start_to_start
        - finish_to_finish
        - start_to_finish
        type: string
      dependsOnTaskId:
        type: string
      taskId:
        type: string
    required:
    - dependencyType
    - dependsOnTaskId
    - taskId
    type: object
  dto.ValidateDependencyResponse:
    properties:
      impact:
        $ref: '#/definitions/dto.ValidationImpact'
      valid:
        type: boolean
      warnings:
        items:
          $ref: '#/definitions/dto.DependencyWarning'
        type: array
      wouldCreateCycle:
        type: boolean
    type: object
  dto.ValidationImpact:
    properties:
      affectedTasks:
        type: integer
      estimatedDelay:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Xephyr Project Management Intelligence API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Xephyr API
  version: "1.0"
paths:
...
  /dependencies:
    post:
      consumes:
      - application/json
      description: Create a new dependency between tasks
      parameters:
      - description: Dependency creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDependencyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateDependencyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create dependency
      tags:
      - dependencies
  /dependencies/{dependencyId}:
    delete:
      consumes:
      - application/json
      description: Remove a dependency
      parameters:
      - description: Dependency ID
        in: path
        name: dependencyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete dependency
      tags:
      - dependencies
  /dependencies/critical-path/{projectId}:
    get:
      consumes:
      - application/json
      description: Get the critical path for a project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CriticalPathResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get critical path
      tags:
      - dependencies
  /dependencies/graph/{projectId}:
    get:
      consumes:
      - application/json
      description: Get the dependency graph for visualization
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DependencyGraphResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get dependency graph
      tags:
      - dependencies
  /dependencies/tasks/{taskId}:
    get:
      consumes:
      - application/json
      description: Get all dependencies for a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Include indirect dependencies
        in: query
        name: includeIndirect
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.TaskDependenciesResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Get task dependencies
      tags:
      - dependencies
  /dependencies/validate:
    post:
      consumes:
      - application/json
      description: Validate a potential dependency without creating it
      parameters:
      - description: Validation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ValidateDependencyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ValidateDependencyResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Validate dependency
      tags:
      - dependencies
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"